Insert a Screening Section

Insert a spectrum screening section in RMarkdown reports. Use this function within a code chunk and make sure to set the chunk option results = 'asis'.

report_Screening(files, meta = NULL, settings = NULL, delim)

Arguments

files
character (required): file path or directory where the spectra files are stored.

meta
list (required): A named list with information on the following elements given in .meta. Consider using .meta as a constructor.

settings
list (required): A named list with information on the following elements given in .settings. Consider using .settings as a constructor.

delim
character (required): Keyword specifying the delimiter insterted before the section. Usually "<br>" or "<hr>" for HTML and "\\newpage" for PDF reports.

Value

Text output via cat.

Examples

# NOT RUN {
report_Screening(files = "~/PATH/TO/FILES/",
                 meta = .meta(person = "Max Mustermann",
                              date = "Yesterday",
                              file_suffix = c("_a", "_b"),
                              file_path = "~/ESR/"),
                 settings = .settings(device = "Bruker ESP300-E",
                                      mod_amp = 0.485,
                                      receiver_gain = 2,
                                      freq = 9.628,
                                      conv_time = 20.48,
                                      time_const = 163.84,
                                      sweep_time = 20.972,
                                      power = 25.3,
                                      center_field = 3430,
                                      sweep_width = 40,
                                      scans = c(10, 20, 30, 40)),
                 delim = "\\newpage")
# }