Insert a Experimental Settings Section

Insert a section on spectrometer settings in RMarkdown reports. Use this function within a code chunk and make sure to set the chunk option results = 'asis'.

report_Settings(x = .settings())

.settings(device = c("Bruker ESP300-E", "Bruker ELEXSYS 500")[1],
  mod_amp = NA, receiver_gain = NA, freq = NA, conv_time = NA,
  time_const = NA, sweep_time = NA, power = NA, center_field = NA,
  sweep_width = NA, scans = NA, file = NA)

Arguments

x
list (required): A named list with information on the following elements given in .settings. Consider using .settings as a constructor.

device
character: Name or model number of the ESR spectrometer.

mod_amp
numeric: Modulation Amplitude

receiver_gain
numeric: Receiver Gain

freq
numeric: Microwave Frequency

conv_time
numeric: Conversion Time

time_const
numeric: Time Constant

sweep_time
numeric: Sweep Time

power
numeric: Microwave Power

center_field
integer: Center Field

sweep_width
integer: Sweep Width

scans
integer: Number of scans

file
character: File name

Value

Text output via cat.

Examples

report_Settings(.settings(device = "Bruker ESP300-E",
                          mod_amp = 0.485,
                          receiver_gain = 2,
                          freq = 9.628,
                          conv_time = 20.48,
                          time_const = 163.84,
                          sweep_time = 20.972,
                          power = 25.3,
                          center_field = 3430,
                          sweep_width = 40,
                          scans = c(10, 20, 30, 40)))

#> + **Device**: Bruker ESP300-E 
#> 
#> + **Modulation Amplitude**: 0.485 
#> 
#> + **Receiver gain**: 2 
#> 
#> + **Microwave Frequency (GHz)**: 9.628 
#> 
#> + **Conversion time (ms)**: 20.48 
#> 
#> + **Time constant (ms)**: 163.84 
#> 
#> + **Sweep time (s)**: 20.972 
#> 
#> + **Microwave power (mW)**: 25.3 
#> 
#> + **Center field (G)**: 3430 
#> 
#> + **Sweep width (G)**: 40 
#> 
#> + **Number of scans**: 10, 20, 30, 40 
#> 
#> + **File(s)**: NA 
#>