Insert a Equivalent Dose Calculation Section

Insert a section on the calculation of the equivalent dose for a given set of input files. Use this function within a code chunk and make sure to set the chunk option results = 'asis'.

report_DE(files, title = "Equivalent dose estimation",
  title.suffix = "", natural = c(1, 2), natural.comment, dose,
  meta = NULL, settings = NULL, fit.repeated = FALSE,
  sample.weight = NULL, interval, th = 10, model = "EXP",
  amplitudes = NULL, spike = NULL, bootstrap = TRUE,
  delim = "\\newpage", ...)

Arguments

files
character (required): file path or directory where the spectra files are stored.

title
character (optional): Main title of the section (defaults to 'Equivalent dose estimation').

title.suffix
character (optional): Optional text to appear in brackets after the title. Useful if the report contains multiple DE measurements.

natural
numeric (optional): Optional vector of integer values specifying which of the spectra are repeated natural dose measurements (defaults to c(1, 2)).

natural.comment
character (optional): An optional comment on the natural dose. Defaults to the statement that repeated natural dose measurements were done first and last in the measurement series.

dose
character or numeric (required): Either a file path to a text file containing the irradiation doses or a numeric vector.

meta
list (required): A named list with information on the following elements given in .meta. Consider using .meta as a constructor.

settings
list (required): A named list with information on the following elements given in .settings. Consider using .settings as a constructor.

fit.repeated
logical (optional): Whether a repeated zero-dose measurement should be considered for fitting or not. If `TRUE` all repeated zero-dose measurements (indicated by argument `natural`) are removed from the data set before fitting.

sample.weight
character or numeric (optional): Either a file path to a text file containing the sample weights or a numeric vector.

interval
numeric (required): A vector of length two specifying the range of x-values where peaks are searched

th
numeric (optional): An integer specifying the number of neighbouring values to compare each x-value to.

model
character (required): Currently implemented models: single-saturating exponential ("EXP"), linear ("LIN").

amplitudes
numeric (optional): A numeric vector of signal amplitudes. If provided, automatic peak finding and amplitude calculation is skipped.

spike
character (optional): Path to a single file of a spike ESR spectrum. The amplitude of the spike signal will be subtracted from all other amplitudes (default TRUE).

bootstrap
logical (optional): Include calculation of the equivalent dose using a bootstrap approach.

delim
character (required): Keyword specifying the delimiter insterted before the section. Usually "<br>" or "<hr>" for HTML and "\\newpage" for PDF reports.

...
further arguments passed to fit_DRC, calc_DePlateau and read_Spectrum.

Value

Text output via cat.

A [list] with all intermediate and final results useful for debugging.

Examples

# not available yet